Use more double variables

+++ model/crystal.hpp
@@ -20,7 +20,7 @@ class Crystal : public Element
 private :
 
     Point center;
-    int radius;
+    double radius;
     int amplifier;
 
 
@@ -34,7 +34,7 @@ public:
      * @param r le rayon du cristal
      * @param m le amplifierificateur de longueur d'onde du cristal
      */
-    Crystal(const Point &, const int, const int);
+    Crystal(const Point &, const double, const int);
 
     /**
      * Retourne la coordonnée du centre du cristal
